clean version setup.py
