+++ README.md
@@ -1,4 +1,6 @@
 # QuestEval
+![GitHub](https://img.shields.io/github/license/recitalAI/QuestEval)
+![PyPI](https://img.shields.io/pypi/v/questeval)
 
 QuestEval is a **NLG metric** to assess if two different inputs contain the same information. The metric, based on Question Generation and Answering can deal with **multimodal** and **multilingual** inputs. 
 It is the result from an (on-going) international collaboration, and so far it tackles various tasks:
@@ -40,14 +42,14 @@ The pre-trained QA/QG models will be automatically downloaded. Alternatively, yo
 The default task is *text2text* and the default language is *en*. It allows to measure the content similarity between any English texts:
 
 ```
-from questeval.questeval_metric import QuestEval
+``from questeval.questeval_metric import QuestEval
 
 source = """After wildfires consumed an entire town, students and teachers who had planned for remote classes found some comfort in staying connected amid the chaos."""
 hypothesis = """Ash fell from an apocalyptic orange sky as Jennifer Willin drove home last week from the only school in tiny Berry Creek, Calif., where she had picked up a pair of Wi-Fi hot spots for her daughters’ remote classes. Hours later, her cellphone erupted with an emergency alert: Evacuate immediately. By the next morning, what one official described as a “massive wall of fire” had swept through the entire Northern California town of about 1,200 people, killing nine residents, including a 16-year-old boy, and destroying the school and almost every home and business. Ms. Willin and her family escaped to a cramped hotel room 60 miles away. In her panic, she had forgotten to grab masks, but she had the hot spots, along with her daughters’ laptops and school books. On Monday, the two girls plan to meet with their teachers on Zoom, seeking some comfort amid the chaos. They’re still able to be in school,” Ms. Willin said, “even though the school burned to the ground.” As the worst wildfire season in decades scorches the West amid a still raging pandemic, families and educators who were already starting the strangest and most challenging school year of their lifetimes have been traumatized all over again. Tens of thousands of people have been forced to flee their homes, with some mourning the loss of their entire communities. But amid the twin disasters, the remote learning preparations that schools made for the coronavirus are providing a strange modicum of stability for teachers and students, letting many stay connected and take comfort in an unexpected form of virtual community."""
 
 questeval = QuestEval()
 score = questeval.compute_all(hypothesis, source)
-print(score['scores'])
+print(score['scores'])``
 ```
 Output:
 ```
