+++ requirements.txt
@@ -1,0 +1,8 @@
+pytest==6.2.4
+transformers==4.6.1
+sentencepiece==0.1.95
+datasets==1.7.0
+bert_score==0.3.9
+Unidecode==1.2.0
+spacy==3.0.6
+en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0.tar.gz
